Conflicts:
	willie/db.py
	willie/modules/adminchannel.py
	willie/modules/clock.py
	willie/modules/rss.py
	willie/tools.py

+++ willie.py
@@ -25,12 +25,13 @@ import os
 import argparse
 import signal
 
-from willie.__init__ import run
+from willie.__init__ import run, __version__
 from willie.config import Config, create_config, ConfigurationError, wizard
 import willie.tools as tools
 import willie.web
 
 homedir = os.path.join(os.path.expanduser('~'), '.willie')
+
 
 def enumerate_configs(extension='.cfg'):
     configfiles = []
@@ -63,33 +64,35 @@ def main(argv=None):
         parser.add_argument('-c', '--config', metavar='filename',
                             help='use a specific configuration file')
         parser.add_argument("-d", '--fork', action="store_true",
-                          dest="deamonize", help="Deamonize willie")
+                            dest="deamonize", help="Deamonize willie")
         parser.add_argument("-q", '--quit', action="store_true", dest="quit",
-                          help="Gracefully quit Willie")
+                            help="Gracefully quit Willie")
         parser.add_argument("-k", '--kill', action="store_true", dest="kill",
-                          help="Kill Willie")
+                            help="Kill Willie")
         parser.add_argument('--exit-on-error', action="store_true",
-                          dest="exit_on_error", help=(
-                              "Exit immediately on every error instead of "
-                              "trying to recover"))
+                            dest="exit_on_error", help=(
+                                "Exit immediately on every error instead of "
+                                "trying to recover"))
         parser.add_argument("-l", '--list', action="store_true",
-                          dest="list_configs",
-                          help="List all config files found")
+                            dest="list_configs",
+                            help="List all config files found")
         parser.add_argument("-m", '--migrate', action="store_true",
-                          dest="migrate_configs",
-                          help="Migrate config files to the new format")
+                            dest="migrate_configs",
+                            help="Migrate config files to the new format")
         parser.add_argument('--quiet', action="store_true", dest="quiet",
-                          help="Supress all output")
+                            help="Supress all output")
         parser.add_argument('-w', '--configure-all', action='store_true',
-                          dest='wizard', help='Run the configuration wizard.')
+                            dest='wizard', help='Run the configuration wizard.')
         parser.add_argument('--configure-modules', action='store_true',
-                          dest='mod_wizard', help=(
-                              'Run the configuration wizard, but only for the '
-                              'module configuration options.'))
+                            dest='mod_wizard', help=(
+                                'Run the configuration wizard, but only for the '
+                                'module configuration options.'))
         parser.add_argument('--configure-database', action='store_true',
-                          dest='db_wizard', help=(
-                              'Run the configuration wizard, but only for the '
-                              'database configuration options.'))
+                            dest='db_wizard', help=(
+                                'Run the configuration wizard, but only for the '
+                                'database configuration options.'))
+        parser.add_argument('-v', '--version', action="store_true",
+                            dest="version", help="Show version number and exit")
         opts = parser.parse_args()
 
         try:
@@ -100,7 +103,14 @@ def main(argv=None):
             # Windows don't have os.getuid/os.geteuid
             pass
 
-        if opts.wizard:
+        if opts.version:
+            py_ver = '%s.%s.%s' % (sys.version_info.major,
+                                   sys.version_info.minor,
+                                   sys.version_info.micro)
+            print('Willie %s (running on python %s)' % (__version__, py_ver))
+            print('http://willie.dftba.net/')
+            return
+        elif opts.wizard:
             wizard('all', opts.config)
             return
         elif opts.mod_wizard:
@@ -160,7 +170,7 @@ def main(argv=None):
         sys.stderr = tools.OutputRedirect(logfile, True, opts.quiet)
         sys.stdout = tools.OutputRedirect(logfile, False, opts.quiet)
 
-        #Handle --quit, --kill and saving the PID to file
+        # Handle --quit, --kill and saving the PID to file
         pid_dir = config_module.core.pid_dir or homedir
         if opts.config is None:
             pid_file_path = os.path.join(pid_dir, 'willie.pid')
@@ -170,10 +180,12 @@ def main(argv=None):
                 basename = basename[:-4]
             pid_file_path = os.path.join(pid_dir, 'willie-%s.pid' % basename)
         if os.path.isfile(pid_file_path):
-            pid_file = open(pid_file_path, 'r')
-            old_pid = int(pid_file.read())
-            pid_file.close()
-            if tools.check_pid(old_pid):
+            with open(pid_file_path, 'r') as pid_file:
+                try:
+                    old_pid = int(pid_file.read())
+                except ValueError:
+                    old_pid = None
+            if old_pid is not None and tools.check_pid(old_pid):
                 if not opts.quit and not opts.kill:
                     stderr('There\'s already a Willie instance running with this config file')
                     stderr('Try using the --quit or the --kill options')
@@ -199,9 +211,8 @@ def main(argv=None):
             child_pid = os.fork()
             if child_pid is not 0:
                 sys.exit()
-        pid_file = open(pid_file_path, 'w')
-        pid_file.write(str(os.getpid()))
-        pid_file.close()
+        with open(pid_file_path, 'w') as pid_file:
+            pid_file.write(str(os.getpid()))
         config_module.pid_file_path = pid_file_path
 
         # Step Five: Initialise And Run willie
